+++ packages/editor/editor-core/src/ui/Addon/ClickAreaBlock/index.tsx
@@ -34,6 +34,7 @@ export default class ClickAreaBlock extends React.Component<Props> {
     const contentArea = event.currentTarget.querySelector(
       '.ak-editor-content-area',
     );
+    const editorFocused = view!.hasFocus();
 
     // @see https://product-fabric.atlassian.net/browse/ED-4287
     // click event gets triggered twice on a checkbox (on <label> first and then on <input>)
@@ -46,7 +47,9 @@ export default class ClickAreaBlock extends React.Component<Props> {
       '[data-editor-popup]',
     );
     if (
-      (!contentArea || !insideContentArea(event.target.parentNode)) &&
+      (!contentArea ||
+        !insideContentArea(event.target.parentNode) ||
+        editorFocused === false) &&
       !isInputClicked &&
       !isPopupClicked &&
       view
@@ -54,7 +57,7 @@ export default class ClickAreaBlock extends React.Component<Props> {
       const { dispatch, dom } = view;
       const bottomAreaClicked =
         event.clientY > dom.getBoundingClientRect().bottom;
-      const isParagrpahAppended = bottomAreaClicked
+      const isParagraphAppended = bottomAreaClicked
         ? createParagraphAtEnd()(view.state, dispatch)
         : false;
       const isGapCursorSet = setCursorForTopLevelBlocks(
@@ -63,7 +66,7 @@ export default class ClickAreaBlock extends React.Component<Props> {
         view.posAtCoords.bind(view),
       )(view.state, dispatch);
 
-      if (isParagrpahAppended || isGapCursorSet) {
+      if (isParagraphAppended || isGapCursorSet) {
         view.focus();
         event.stopPropagation();
       }