+++ packages/navigation/centralised-app-switcher/src/primitives/manage-button.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default () => <button>Manage</button>;