+++ packages/navigation/centralised-app-switcher/src/providers/mock-provider.ts
@@ -1,0 +1,7 @@
+import asDataProvider from './as-data-provider';
+
+const SOME_STATIC_DATA = {
+  data: 'yay!',
+};
+
+export default asDataProvider(() => SOME_STATIC_DATA);