+++ packages/navigation/centralised-app-switcher/src/utils/fetch.ts
@@ -1,0 +1,1 @@
+export const fetchJson = url => fetch(url).then(response => response.json());