+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @atlaskit/renderer
 
+## 35.0.4
+- [patch] [dfdfaf2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dfdfaf2):
+
+  - ED-5493 Fix Media Cards bigger in renderer than editor
+
+## 35.0.3
+- [patch] [48abc90](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/48abc90):
+
+  - Fixed date border radius in renderer
+
+## 35.0.2
+- [patch] [609d32d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/609d32d):
+
+  - ED-5966 Fix issue where renderer incorrectly displayed numbered tables without headers
+
+## 35.0.1
+- Updated dependencies [0c116d6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0c116d6):
+  - @atlaskit/editor-json-transformer@4.1.5
+  - @atlaskit/editor-test-helpers@6.3.8
+  - @atlaskit/editor-common@28.0.2
+  - @atlaskit/util-data-test@10.0.32
+  - @atlaskit/mention@16.0.0
+
 ## 35.0.0
 - Updated dependencies [cbb8cb5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/cbb8cb5):
   - @atlaskit/editor-common@28.0.0